pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eodash_catalog"
dynamic = ["version"]
description = "This package is intended to help create a compatible STAC catalog for the eodash dashboard client. It supports configuration of multiple endpoint types for information extraction."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [{ name = "Daniel Santillan", email = "daniel.santillan@eox.at" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "setuptools<71",
  "oauthlib<3.3",
  "requests-oauthlib<1.3.2",
  "python-dotenv<1.1.0",
  "pystac-client<1",
  "pyyaml<7",
  "click<9",
  "requests<3",
  "python-dateutil<3",
  "structlog<22.0",
  "OWSLib",
  "spdx-lookup<=0.3.3",
  "pystac[validation]<2",
  "stac_geoparquet<=0.7.0"
]
[project.scripts]
eodash_catalog = "eodash_catalog.generate_indicators:process_catalogs"
[project.urls]
Documentation = "https://github.com/eodash/eodash_catalog#readme"
Issues = "https://github.com/eodash/eodash_catalog/issues"
Source = "https://github.com/eodash/eodash_catalog"
[project.optional-dependencies]


[tool.hatch.version]
path = "src/eodash_catalog/__about__.py"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]
[tool.hatch.metadata]
allow-direct-references = true
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/eodash_catalog tests}"

[tool.coverage.run]
source_pkgs = ["eodash_catalog", "tests"]
branch = true
parallel = true
omit = ["src/eodash_catalog/__about__.py"]

[tool.coverage.paths]
eodash_catalog = ["src/eodash_catalog", "*/eodash_catalog/src/eodash_catalog"]
tests = ["tests", "*/eodash_catalog/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.mypy]
disable_error_code = ["import-untyped"]
